Fix wrong matches for event handlers
We have many event handlers defined as with a wildcard parametrized type
like `on(DeviceUpdatedEvent<?> event)`. The plugin was considering that
as a handler for `java.lang.Object` because it was taking the `?` as the
type. Since it's the superclass for everything, all those handlers were
always showing, making it impossible to find the real ones in that huge
list.

Helped by Codex. Not completely sure about some parts of the code. Will
have to analyze more at some point and do a real PR and ticket to their
repo.